// frontend-v2/src/interfaces/ResponseInterface.ts
// Courses
export interface ICourse {
  courseCode: string;
  archived: boolean;
  attributes: string[];
  calendar: string;
  campus: string;
  description: string;
  enrolmentRules: string;
  equivalents: string[];
  exclusions: string[];
  faculty: string;
  fieldOfEducation: string;
  genEd: boolean;
  level: number;
  school: string;
  studyLevel: string;
  terms: number[];
  title: string;
  uoc: number;
  rating: number; // same as overallRating
  reviewCount: number;
  overallRating: number;
  manageability: number;
  usefulness: number;
  enjoyability: number;
}

// Reviews
export interface IReview {
  reviewId: string;
  zid: string;
  courseCode: string;
  authorName: string;
  description: string;
  grade: number;
  termTaken: string;
  createdTimestamp: string;
  updatedTimestamp: string;
  upvotes: string[];
  manageability: number;
  enjoyability: number;
  usefulness: number;
  overallRating: number;
}

// Reports
export enum ReportStatus {
  UNSEEN = "UNSEEN",
  SEEN = "SEEN",
  REMOVED = "REMOVED",
  SETTLED = "SETTLED",
}

export interface IReport {
  reportId: string;
  review: IReview;
  zid: string;
  status: ReportStatus;
  reason: string;
  createdTimestamp: Date;
  updatedTimestamp: Date;
}

// Users
export interface IUser {
  zid: string;
  isAdmin: boolean;
  bookmarkedCourses: ICourse[];
  bookmarkedReviews: IReview[];
  reports: IReport[];
  reviews: IReview[];
}

// Tokens
export interface ITokenData {
  expiresIn: string;
  token: string;
}

export interface IToken {
  zid: string;
}

// API Responses

// Errors
export interface ApiError {
  errorCode: string;
  errorMessage: string;
}

export interface IPostUserResponse {
  user: IUser;
  token: ITokenData;
}

export interface IGetCoursesResponse {
  courses: ICourse[];
}

export interface IPostReviewRequestBody {
  zid: string;
  courseCode: string;
  authorName: string;
  title: string;
  description: string;
  grade: number;
  termTaken: string;
  manageability: number;
  usefulness: number;
  enjoyability: number;
  overallRating: number;
}

export interface IPostReviewResponse {
  review: IReview;
}
export interface IGetReviewsResponse {
  reviews: IReview[];
}